fixed define_scope macro
