+++ leptos_i18n_codegen/src/utils/mod.rs
@@ -1,48 +1,5 @@
 use proc_macro2::TokenStream;
-use quote::{ToTokens, TokenStreamExt, format_ident, quote};
-use syn::Token;
-
-pub enum Keys {
-    SingleKey(syn::Ident),
-    Subkeys(Vec<syn::Ident>),
-}
-
-fn parse_subkeys(input: syn::parse::ParseStream, keys: &mut Vec<syn::Ident>) -> syn::Result<()> {
-    keys.push(input.parse()?);
-    while input.peek(Token![.]) {
-        input.parse::<Token![.]>()?;
-        keys.push(input.parse()?);
-    }
-    Ok(())
-}
-
-impl syn::parse::Parse for Keys {
-    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
-        let first_key = input.parse()?;
-        let dot = input.peek(Token![.]);
-        if dot || input.peek(Token![::]) {
-            if dot {
-                input.parse::<Token![.]>()?;
-            } else {
-                input.parse::<Token![::]>()?;
-            }
-            let mut keys = vec![first_key];
-            parse_subkeys(input, &mut keys)?;
-            Ok(Keys::Subkeys(keys))
-        } else {
-            Ok(Keys::SingleKey(first_key))
-        }
-    }
-}
-
-impl quote::ToTokens for Keys {
-    fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
-        match self {
-            Keys::SingleKey(key) => tokens.append(key.clone()),
-            Keys::Subkeys(keys) => tokens.append_separated(keys, quote!(().)),
-        }
-    }
-}
+use quote::{ToTokens, format_ident, quote};
 
 #[derive(Debug, Clone)]
 pub enum EitherOfWrapper {
@@ -53,10 +10,11 @@ pub enum EitherOfWrapper {
 }
 
 impl EitherOfWrapper {
+    #[track_caller]
     pub fn new(size: usize) -> EitherOfWrapper {
         match size {
             0 => {
-                unreachable!("0 locales ? how is this possible ? should have been checked by now.")
+                unreachable!("EitherOfWrapper requires at least one element.")
             }
             1 => EitherOfWrapper::Single,
             2 => EitherOfWrapper::Duo,
